==> End-to-End Pipeline/2. cleaning categorical features.ipynb <==
## Pipeline: Clean Categorical Features

Using the Titanic dataset from [this](https://www.kaggle.com/c/titanic/overview) Kaggle competition.

This dataset contains information about 891 people who were on board the ship when departed on April 15th, 1912. As noted in the description on Kaggle's website, some people aboard the ship were more likely to survive the wreck than others. There were not enough lifeboats for everybody so women, children, and the upper-class were prioritized. Using the information about these 891 passengers, the challenge is to build a model to predict which people would survive based on the following fields:

- **Name** (str) - Name of the passenger
- **Pclass** (int) - Ticket class
- **Sex** (str) - Sex of the passenger
- **Age** (float) - Age in years
- **SibSp** (int) - Number of siblings and spouses aboard
- **Parch** (int) - Number of parents and children aboard
- **Ticket** (str) - Ticket number
- **Fare** (float) - Passenger fare
- **Cabin** (str) - Cabin number
- **Embarked** (str) - Port of embarkation (C = Cherbourg, Q = Queenstown, S = Southampton)

**This notebook will implement some of the cleaning that was done in Section 2: EDA & Data Cleaning**



### Read in Data

import numpy as np
import pandas as pd

from google.colab import drive
drive.mount('/gdrive')
%cd/gdrive

# Mounting google drive to access data
from google.colab import drive
drive.mount('/content/drive')

# Importing the libraries

import pandas as pd
from sklearn.model_selection import train_test_split

# Read csv files
titanic_data= '/content/drive/MyDrive/Summer_Learning/Ex_Files_Applied_Machine_Learning/Exercise_Files/titanic.csv'
titanic = pd.read_csv(titanic_data)

titanic.head(5)

### Clean categorical variables
1. Create an indicator variable in place of `Cabin`
2. Convert `Sex` to a numeric variable
3. Drop irrelevant/repetitive variables (`Cabin`, `Embarked`, `Name`, `Ticket`)

#### Create indicator for `Cabin`

titanic['Cabin_ind'] = np.where(titanic['Cabin'].isnull(), 0, 1)

#### Convert `Sex` to numeric

gender_num = {'male': 0, 'female': 1}

titanic['Sex'] = titanic['Sex'].map(gender_num)

#### Drop unnecessary variables

titanic.drop(['Cabin', 'Embarked', 'Name', 'Ticket'], axis=1, inplace=True)
titanic.head()

### Write out cleaned data

titanic.to_csv('/content/drive/MyDrive/Summer_Learning/Ex_Files_Applied_Machine_Learning/Exercise_Files/titanic_cleaned1.csv', index=False)
