==> End-to-End Pipeline/3. Splitting data into train, val, test set.ipynb <==
## Pipeline: Split data into train, validation, and test set

Using the Titanic dataset from [this](https://www.kaggle.com/c/titanic/overview) Kaggle competition.

In this section, we will split the data into train, validation, and test set in preparation for fitting a basic model in the next section.

### Read in Data

import pandas as pd
from sklearn.model_selection import train_test_split

# Mounting google drive to access data
from google.colab import drive
drive.mount('/content/drive')

from google.colab import drive
drive.mount('/gdrive')



# Read csv files
titanic_data= '/content/drive/MyDrive/Summer_Learning/Ex_Files_Applied_Machine_Learning/Exercise_Files/titanic_cleaned.csv'
titanic = pd.read_csv(titanic_data)

titanic.head(5)

### Split into train, validation, and test set


features = titanic.drop('Survived', axis=1)
labels = titanic['Survived']

X_train, X_test, y_train, y_test = train_test_split(features, labels, test_size=0.4, random_state=42)
X_test, X_val, y_test, y_val = train_test_split(X_test, y_test, test_size=0.5, random_state=42)

for dataset in [y_train, y_val, y_test]:
  # round percentage to two decimal places
    print(round(len(dataset) / len(labels), 2))

### Write out data

X_train.to_csv('/content/drive/MyDrive/Summer_Learning/Ex_Files_Applied_Machine_Learning/Exercise_Files/train_features.csv', index=False)
X_val.to_csv('/content/drive/MyDrive/Summer_Learning/Ex_Files_Applied_Machine_Learning/Exercise_Files/val_features.csv', index=False)
X_test.to_csv('/content/drive/MyDrive/Summer_Learning/Ex_Files_Applied_Machine_Learning/Exercise_Files/test_features.csv', index=False)

y_train.to_csv('/content/drive/MyDrive/Summer_Learning/Ex_Files_Applied_Machine_Learning/Exercise_Files/train_labels.csv', index=False)
y_val.to_csv('/content/drive/MyDrive/Summer_Learning/Ex_Files_Applied_Machine_Learning/Exercise_Files/val_labels.csv', index=False)
y_test.to_csv('/content/drive/MyDrive/Summer_Learning/Ex_Files_Applied_Machine_Learning/Exercise_Files/test_labels.csv', index=False)
