==> End-to-End Pipeline/5. Hyperparameter Tuning.ipynb <==
## Pipeline: Tune hyperparameters

Using the Titanic dataset from [this](https://www.kaggle.com/c/titanic/overview) Kaggle competition.

we will tune the hyperparameters for the basic model we fit in the last section.

### Read in data



from google.colab import drive
drive.mount('/gdrive')
%cd/gdrive

# Mounting google drive to access data
from google.colab import drive
drive.mount('/content/drive')

import pandas as pd
from sklearn.ensemble import RandomForestClassifier
# Allows grid search inside the cross-validation
from sklearn.model_selection import GridSearchCV

tr_features = pd.read_csv('/content/drive/MyDrive/Summer_Learning/Ex_Files_Applied_Machine_Learning/Exercise_Files/train_features.csv')
tr_labels = pd.read_csv('/content/drive/MyDrive/Summer_Learning/Ex_Files_Applied_Machine_Learning/Exercise_Files/train_labels.csv', header=None)

### Hyperparameter tuning


def print_results(results):
    print('BEST PARAMS: {}\n'.format(results.best_params_))

    means = results.cv_results_['mean_test_score']
    stds = results.cv_results_['std_test_score']
    for mean, std, params in zip(means, stds, results.cv_results_['params']):
        print('{} (+/-{}) for {}'.format(round(mean, 3), round(std * 2, 3), params))

rf = RandomForestClassifier()
parameters = {
    'n_estimators': [5, 50, 100],
    'max_depth': [2, 10, 20, None]
}

cv = GridSearchCV(rf, parameters, cv=5)
cv.fit(tr_features, tr_labels.values.ravel())

print_results(cv)
